添加门店接口调试

+++ src/views/dealersMgt/dealerMgt/index.vue
@@ -18,20 +18,20 @@
       <right-toolbar :showSearch.sync="showSearch" @queryTable="getList"></right-toolbar>
     </el-row>
 
-    <el-table height="600" v-if="refreshTable" v-loading="loading" :data="deptList" row-key="deptId"
+    <el-table height="600" size="medium" v-if="refreshTable" v-loading="loading" :data="deptList" row-key="deptId"
       :default-expand-all="isExpandAll" :tree-props="{ children: 'children', hasChildren: 'hasChildren' }">
       <el-table-column prop="name" width="180" fixed label="经销商名称"></el-table-column>
       <el-table-column prop="contactName" label="联系人" fixed width="100" align="center"></el-table-column>
       <el-table-column prop="contactTel" label="联系电话" fixed width="140">
       </el-table-column>
       <el-table-column label="银行账号" prop="bankAccount" width="200"> </el-table-column>
-      <el-table-column prop="bankAccountName" label="银行开户名" align="center" width="100"></el-table-column>
+      <el-table-column prop="bankAccountName" label="银行开户名"  width="100"></el-table-column>
       <el-table-column prop="bankName" label="开户银行"></el-table-column>
       <el-table-column prop="g" label="云通道" width="100" align="center"></el-table-column>
       <el-table-column prop="strategyId" label="经销商账号" width="150"></el-table-column>
       <el-table-column label="门店剩余数量" align="center" width="100" class-name="small-padding fixed-width">
         <template slot-scope="scope">
-          <el-button size="mini" type="text" @click="dealerBalance()">查看</el-button>
+          <el-button size="mini" type="text" @click="dealerBalance(scope.row)">查看</el-button>
         </template>
       </el-table-column>
       <el-table-column prop="remark" label="备注"></el-table-column>
@@ -61,47 +61,47 @@
       <el-form ref="openForm" :model="openForm" :rules="rules" label-width="100px" style="padding-left: 29px">
         <el-row>
           <el-col :span="22">
-            <el-form-item label="经销商名称 :">
+            <el-form-item label="经销商名称 :" prop="name">
               <el-input v-model="openForm.name" placeholder="请输入经销商名称" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="联系人 :">
+            <el-form-item label="联系人 :" prop="contactName">
               <el-input v-model="openForm.contactName" placeholder="请输入联系人" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="联系电话 :" prop="phone">
+            <el-form-item label="联系电话 :" prop="contactTel">
               <el-input v-model="openForm.contactTel" placeholder="请输入联系电话" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="银行账号 :">
+            <el-form-item label="银行账号 :" prop="bankAccount">
               <el-input v-model="openForm.bankAccount" placeholder="请输入银行账号" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="银行开户名 :">
+            <el-form-item label="银行开户名 :" prop="bankAccountName">
               <el-input v-model="openForm.bankAccountName" placeholder="请输入银行开户名" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="开户银行 :">
+            <el-form-item label="开户银行 :" prop="bankName">
               <el-input v-model="openForm.bankName" placeholder="请输入开户银行" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="经销商账号 :">
+            <el-form-item label="经销商账号 :" prop="strategyId">
               <el-input v-model="openForm.strategyId" placeholder="请输入经销商账号" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="门店数量 :">
-              <el-input v-model="openForm.h" placeholder="请输入门店数量" />
+            <el-form-item label="门店数量 :" prop="quantity">
+              <el-input v-model="openForm.quantity" placeholder="请输入门店数量" />
             </el-form-item>
           </el-col>
           <el-col :span="22">
-            <el-form-item label="备注 :">
+            <el-form-item label="备注 :" prop="remark">
               <el-input type="textarea" v-model="openForm.remark" placeholder="请输入备注" />
             </el-form-item>
           </el-col>
@@ -113,7 +113,7 @@
     </el-dialog>
     <!-- 添加设置通道 -->
     <el-dialog title="设备通道" :visible.sync="openPassage" width="600px" @close="handleClose" append-to-body>
-      <el-form ref="form" :model="passageForm" :rules="rules" style="padding-left: 29px" label-width="100px">
+      <el-form ref="form" :model="passageForm"  style="padding-left: 29px" label-width="100px">
         <el-row>
           <el-col :span="22">
             <el-form-item label="经销商名称 :">
@@ -139,7 +139,7 @@
     </el-dialog>
     <!-- 查看经销商余额 -->
     <el-dialog title="经销商余额" @close="handleClose" :visible.sync="openDealer" width="550px" append-to-body>
-      <el-form ref="form" :model="dealerForm" :rules="rules" label-width="100px" style="padding-left: 29px">
+      <el-form ref="form" :model="dealerForm"  label-width="100px" style="padding-left: 29px">
         <el-row>
           <el-col :span="22">
             <el-form-item label="经销商名称">
@@ -194,7 +194,7 @@ export default {
       queryParams: {
         pageIndex: 1,
         size: 10,
-        parameter: "",
+        parameter: undefined,
       },
       // 维护经销商表单参数
       openForm: {},
@@ -204,13 +204,21 @@ export default {
       dealerForm: {},
       // 表单校验
       rules: {
-        phone: [
-          {
+         name: [{ required: true, message: "名称不能为空", trigger: "blur" },],
+         contactName: [{ required: true, message: "联系人不能为空", trigger: "blur" },],
+        contactTel: [
+          { required: true,
             pattern: /^1[3|4|5|6|7|8|9][0-9]\d{8}$/,
             message: "请输入正确的手机号码",
             trigger: "blur",
           },
         ],
+         bankAccount: [{ required: true,type: 'number', message: "请输入正确的银行账户", trigger: "blur" },],
+         bankAccountName: [{ required: true, message: "开户名不能为空", trigger: "blur" },],
+         bankName: [{ required: true, message: "开户银行不能为空", trigger: "blur" },],
+         strategyId: [{ required: true, message: "名称不能为空", trigger: "blur" },],
+         quantity: [{ required: false, message: "经销商账号不能为空", trigger: "blur" },],
+         remark: [{ required: false},],
       },
     };
   },
@@ -221,22 +229,28 @@ export default {
     /** 查询经销商列表 */
     getList() {
       this.loading = true;
-      var indexPage=0
-      var parameter=this.queryParams.parameter
-      if (this.queryParams.parameter == ''||this.queryParams.parameter == null||this.queryParams.parameter == undefined) {
+      var indexPage = 0
+      if (this.queryParams.parameter == undefined) {
         indexPage = this.queryParams.pageIndex - 1
-      } 
+      }
       const queryParams = {
         pageIndex: indexPage,
         size: this.queryParams.size,
-        parameter: parameter,
+        parameter: this.queryParams.parameter,
       }
       pageAgency(queryParams).then(res => {
-        this.deptList = res.result.data
-        this.total = parseInt(res.result.items)
-        this.loading = false;
-      });
-
+        if (res.type == "success" && res.code == 200) {
+          this.deptList = res.result.data
+          this.total = parseInt(res.result.items)
+          this.loading = false;
+        } else {
+          this.$message.error('获取数据失败,请重试');
+        }
+      }).catch((err) => {
+        this.$notify.error({
+          title: err
+        });
+      })
     },
 
     /** 重置按钮操作 */
@@ -279,21 +293,29 @@ export default {
             this.getList();
             this.openForm = {};
           } else {
-            this.$message.warning('新增失败,请重试');
+            this.$message.error('新增失败,请重试');
             this.getList();
           }
-        });
+        }).catch((err) => {
+          this.$notify.error({
+            title: err
+          });
+        })
       } else {
         this.openForm.userName = "13310249009"
         editsAgency(this.openForm).then(res => {
           if (res.type == "success" && res.code == 200) {
             this.$message.success('修改成功');
             this.getList();
-          }else {
-            this.$message.warning('修改失败,请重试');
+          } else {
+            this.$message.error('修改失败,请重试');
             this.getList();
           }
-        });
+        }).catch((err) => {
+          this.$notify.error({
+            title: err
+          });
+        })
       }
 
     },
@@ -338,16 +360,33 @@ export default {
     },
     // 删除按钮
     deleteRow(row) {
-      deleteAgency(row.id).then(res => {
-        if (res.type == 'success' && res.code == 200) {
-          this.$message.success('删除成功');
-          this.getList();
-        } else {
-          this.$message.warning('删除失败,请重试');
-        }
-      });
+      this.$confirm("此操作将永久删除该数据, 是否继续?", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning",
+      })
+        .then(() => {
+          deleteAgency(row.id).then(res => {
+            if (res.type == 'success' && res.code == 200) {
+              this.$message.success('删除成功');
+              this.getList();
+            } else {
+              this.$message.warning('删除失败,请重试');
+            }
+          }).catch((err) => {
+          this.$notify.error({
+          title: err
+        });
+        })
+        }).catch(() => {
+          /* this.$message.info({
+            type: "info",
+            message: "已取消删除",
+          }); */
+        });
+
     },
-    
+
     // 锁定按钮
     lockBtn(row) {
       console.log(row)
